fixed a few examples

+++ Examples/src/Templates/Fibonacci/fibonacci.hpp
@@ -29,9 +29,9 @@ template<> struct fib<0>{
   @note beware of integer underflow if you use unsigned int. Here there is no problem
   but in general make sure that you will never compute 0u -1
 */
-constexpr unsigned int Fibonacci(unsigned int N)
+constexpr unsigned long int Fibonacci(unsigned long int N)
 {
-  return N <=2u ? 1u: Fibonacci(N-1)+Fibonacci(N-2);
+  return N <=2ul ? 1ul: Fibonacci(N-1)+Fibonacci(N-2);
 }
